requirements.txt
amqp==2.5.2
asgiref==3.2.7
beautifulsoup4==4.9.0
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.6
django-celery-results==1.2.1
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
drf-yasg==1.17.1
flex==6.14.1
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonpointer==2.0
kombu==4.6.8
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.3
psycopg2-binary==2.8.5
Pygments==2.6.1
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.25.9
validate-email==1.3
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
